server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
fasttext-wheel==0.9.2
filelock==3.13.1
fsspec==2023.12.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
iniconfig==2.0.0
numpy==1.26.2
packaging==23.2
pluggy==1.3.0
pybind11==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
